refactor to share image client code between collector and eraser
