+++ pkg/utils/utils.go
@@ -1,0 +1,92 @@
+package pkg
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net"
+	"net/url"
+	"time"
+
+	"google.golang.org/grpc"
+	pb "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
+
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+const (
+	// unixProtocol is the network protocol of unix socket.
+	unixProtocol = "unix"
+)
+
+var (
+	// Timeout  of connecting to server (default: 5m).
+	timeout                  = 5 * time.Minute
+	errProtocolNotSupported  = errors.New("protocol not supported")
+	errEndpointDeprecated    = errors.New("endpoint is deprecated, please consider using full url format")
+	errOnlySupportUnixSocket = errors.New("only support unix socket endpoint")
+	log                      = logf.Log.WithName("eraser")
+)
+
+func GetAddressAndDialer(endpoint string) (string, func(ctx context.Context, addr string) (net.Conn, error), error) {
+	protocol, addr, err := ParseEndpointWithFallbackProtocol(endpoint, unixProtocol)
+	if err != nil {
+		return "", nil, err
+	}
+	if protocol != unixProtocol {
+		return "", nil, errOnlySupportUnixSocket
+	}
+
+	return addr, dial, nil
+}
+
+func dial(ctx context.Context, addr string) (net.Conn, error) {
+	return (&net.Dialer{}).DialContext(ctx, unixProtocol, addr)
+}
+
+func ParseEndpointWithFallbackProtocol(endpoint string, fallbackProtocol string) (protocol string, addr string, err error) {
+	if protocol, addr, err = ParseEndpoint(endpoint); err != nil && protocol == "" {
+		fallbackEndpoint := fallbackProtocol + "://" + endpoint
+		protocol, addr, err = ParseEndpoint(fallbackEndpoint)
+		if err != nil {
+			return "", "", err
+		}
+	}
+	return protocol, addr, err
+}
+
+func ParseEndpoint(endpoint string) (string, string, error) {
+	u, err := url.Parse(endpoint)
+	if err != nil {
+		return "", "", fmt.Errorf("error while parsing: %w", err)
+	}
+
+	switch u.Scheme {
+	case "tcp":
+		return "tcp", u.Host, nil
+	case "unix":
+		return "unix", u.Path, nil
+
+	case "":
+		return "", "", fmt.Errorf("using %q as %w", endpoint, errEndpointDeprecated)
+
+	default:
+		return u.Scheme, "", fmt.Errorf("%q: %w", u.Scheme, errProtocolNotSupported)
+	}
+}
+
+func GetImageClient(ctx context.Context, socketPath string) (pb.ImageServiceClient, *grpc.ClientConn, error) {
+	addr, dialer, err := GetAddressAndDialer(socketPath)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	conn, err := grpc.DialContext(ctx, addr, grpc.WithBlock(), grpc.WithInsecure(), grpc.WithContextDialer(dialer))
+	if err != nil {
+		return nil, nil, err
+	}
+
+	imageClient := pb.NewImageServiceClient(conn)
+
+	return imageClient, conn, nil
+}
